done changes in download, upload, term

+++ src/term/controller.ts
@@ -48,33 +48,34 @@ export async function term2(req: Request, res: Response) {
   }
 }
 
-
 export async function promote(req: Request, res: Response) {
   try {
-    const branch = req.body.branchInToken;
-    await dbQuery(` 
-              DELETE FROM studentinfo
-              WHERE sem = 8 where branch = ${branch};
+    const { branchInToken: branch, sem } = req.body;
 
-              UPDATE studentinfo
-                 SET sem = CASE 
-                      WHEN sem = 1 THEN 2
-                      WHEN sem = 2 THEN 3
-                      WHEN sem = 3 THEN 4
-                      WHEN sem = 4 THEN 5
-                      WHEN sem = 5 THEN 6
-                      WHEN sem = 6 THEN 7
-                      WHEN sem = 7 THEN 8
-                  END
-                WHERE sem BETWEEN 1 AND 7 AND branch = ${branch};
+    if (sem === undefined) {
+      const result: any = await dbQuery(`
+        SELECT DISTINCT sem FROM studentinfo WHERE branch = '${branch}';
+      `);
 
-                UPDATE studentinfo
-                  SET token = "undone" where branch = ${branch};
+      const semesters = result.map((row: { sem: number }) => row.sem);
+      return res.json({ done: true, semesters: semesters });
+    } else {
+      await dbQuery(`
+        DELETE FROM studentinfo
+        WHERE sem = 8 AND branch = '${branch}';
 
-                UPDATE COUNTTERM SET COUNT = 1;
+        UPDATE studentinfo
+        SET sem = sem + 1
+        WHERE sem = ${sem} AND branch = '${branch}';
 
-                `);
-    return res.json({ done: true });
+        UPDATE studentinfo
+        SET token = 'undone' WHERE branch = '${branch}';
+
+        `);
+        // UPDATE COUNTTERM SET COUNT = 1;
+
+      return res.json({ done: true });
+    }
   } catch (error) {
     console.error('Error executing query:', error);
     res.status(500).send('Error executing query');
